fix pasteboard for iOS 9 or below
As WhatsApp is using -[UIPasteboard dataForPasteboardType:] for reading data from pasteboard (in -[WhatsAppAppDelegate processThirdPartyStickerPackWithBundleID:]).

Before iOS 10, it should be using -[UIPasteboard setData:forPasteboardType:] instead of -[UIPasteboard addItems:] to avoid keep importing the same sticker packs or pop up "There's a problem with this sticker pack and it can't be added to WhatsApp" message.